firestore-to-google-sheets-pipeline/requirements.txt
functions-framework==3.5.0
google-events==0.11.0
google-api-core==2.15.0
protobuf==4.25.2
cloudevents==1.10.1
google-cloud-firestore==2.14.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.115.0
pytz==2023.3.post1
# Some packages are left for further optimizations, e.g. trigger event imlementation in the database, etc.